re-organize legacy
