--- classylss/tools.py
@@ -1,3 +1,0 @@
-from .gcl import pk_to_xi
-from .gcl import xi_to_pk
-from .gcl import IntegrationMethods
